requirements.txt
Flask 
google-cloud-storage
google-cloud-firestore